refactor: rename proxyutils to forwarder
